pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "deepeval>=3.9.7",
    "fastapi>=0.136.0",
    "fastmcp>=3.2.4",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.8",
    "langsmith>=0.7.32",
    "networkx>=3.6.1",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "pyvis>=0.3.2",
    "pyyaml>=6.0.3",
    "redis>=7.4.0",
    "reportlab>=4.4.10",
    "streamlit>=1.56.0",
    "uvicorn[standard]>=0.44.0",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "mypy>=1.20.1",
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
    "types-PyYAML>=6.0.12",
    "uv-secure>=0.17.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agent", "mcp_servers", "api", "config", "utils", "app"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
exclude = ["tests/", ".venv/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"
markers = [
    "integration: marks tests as integration tests (may require external services)",
]

[tool.coverage.run]
source = ["agent", "mcp_servers", "api", "config", "utils"]
omit = ["tests/*", ".venv/*"]
